rename : leanSlice 위치 변경
