+++ client/src/redux/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import userSlice from "./features/user/userSlice";
-import learnSlice from "./features/user/learnSlice";
+import learnSlice from "./features/learn/learnSlice";
 
 export default configureStore({
   reducer: {
